Split pipeline chain implementations
